+++ Sources/SpeziMedication/Medication/MedicationInstance.swift
@@ -52,4 +52,3 @@ extension MedicationInstance {
         hasher.combine(id)
     }
 }
-
