+++ Sources/SpeziMedication/Resources/Localizable.xcstrings
@@ -53,13 +53,7 @@
     "Every Day" : {
 
     },
-    "Every Few Days" : {
-
-    },
     "Every other Day" : {
-
-    },
-    "Every other day, Every 3 days" : {
 
     },
     "gel" : {
@@ -87,12 +81,6 @@
       }
     },
     "Medication" : {
-
-    },
-    "On Mondays, On Weekdays" : {
-
-    },
-    "On Specific Days of the Week" : {
 
     },
     "Skipped" : {
